# Optimize polynomial evaluation of coefficients in Gauss-Laguerre