I noticed that some of the coefficients could be computed significantly faster using Horner's scheme and combining evaluations to reuse as many computations as possible. This significantly reduces the computation time of these evaluations.

```julia
# Master

julia> @benchmark gl_wrapper($1.2, x, 1.2) setup=(x=rand())
BenchmarkTools.Trial: 10000 samples with 10 evaluations.
 Range (min … max):  1.304 μs …  2.617 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     1.317 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   1.324 μs ± 47.570 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

    ▃ ▇ █ ▆  ▄ ▃              ▁  ▃ ▁                         ▂
  ▇▁█▁█▁█▁█▁▁█▁█▁█▁▇▁▁▇▁█▁▇▁█▁██▁█▁█▁█▁▇▁▁▇▁▇▁▆▁▆▁▁▆▁▆▁▅▁▅▁▄ █
  1.3 μs       Histogram: log(frequency) by time     1.41 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.

# This PR
julia> @benchmark FastGaussQuadrature.gl_bulk($1.2, x, 1.2) setup=(x=rand())
BenchmarkTools.Trial: 10000 samples with 990 evaluations.
 Range (min … max):  44.276 ns … 63.510 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     44.444 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   44.598 ns ±  0.824 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

   ▅█▆▄▃                       ▁                    ▁         ▁
  ██████▆▆▆▅▅▄▄▅▆▅▅▆▆▅▄▆▅▅▆▅▆▅▆██▇▇▅▆▄▄▅▄▄▄▃▁▃▁▃▄▄▆██▅▄▆▅▄▅▁▃ █
  44.3 ns      Histogram: log(frequency) by time      47.7 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

So the new piece of code is over 25x faster. Of course that is just for that piece of code which is a fraction of the whole routine. For the larger function that is called with 

```julia
# Master
julia> @benchmark FastGaussQuadrature.gausslaguerre_asy_bulk(100, 1.0, 99, x, 11) setup=(x=rand()*0.1)
BenchmarkTools.Trial: 10000 samples with 10 evaluations.
 Range (min … max):  1.142 μs …  2.492 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     1.262 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   1.259 μs ± 40.837 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

                          ▄ ▅▁     ▆█▃▃▂                      
  ▂▂▂▂▂▂▁▂▂▂▂▃▁▃▃▃▃▃▂▁▃▃▄▆█▁██▇▆▅▆▁█████▁▇▅▄▄▅▇▁▆▄▃▃▃▁▃▂▂▃▃▂ ▃
  1.14 μs        Histogram: frequency by time        1.35 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.

# This PR
julia> @benchmark FastGaussQuadrature.gausslaguerre_asy_bulk(100, 1.0, 99, x, 11) setup=(x=rand()*0.1)
BenchmarkTools.Trial: 10000 samples with 636 evaluations.
 Range (min … max):  192.151 ns … 391.313 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     220.715 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   216.940 ns ±  22.124 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █  ▃ ▁              █▆ ▄▂▁▁           ▃▄ ▅▁ ▁              ▂▁ ▂
  ██▆███▅▄▃▃▅▄▄▃▁▄▃▃▁▃███████▇▅▅▄▅▆▁▃▅▄▁█████▆█▇▇▆▅▃▁▃▄▄▄▁▃▄▁██ █
  192 ns        Histogram: log(frequency) by time        277 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

So that is roughly 6x faster which is smaller but still very significant. Now calling the larger function...
```julia
# Master
julia> @benchmark FastGaussQuadrature.gausslaguerre_asy(100, x) setup=(x=rand()*0.1)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  162.083 μs … 218.042 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     162.875 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   164.613 μs ±   4.015 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▃▇█▅▃▁▁  ▂▄▂▁   ▁▁ ▁▁▁▁▁▁▂▁▁                                  ▁
  █████████████████████████████▇▇▇▇▆▆▆▆▆▆▄▆▅▇▅▆▅▅▅▆▆▆▅▅▄▅▆▅▆▆▅▆ █
  162 μs        Histogram: log(frequency) by time        182 μs <

 Memory estimate: 3.33 KiB, allocs estimate: 34.

# this PR
julia> @benchmark FastGaussQuadrature.gausslaguerre_asy(100, x) setup=(x=rand()*0.1)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  50.667 μs … 104.958 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     51.209 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   51.948 μs ±   2.506 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▂██▅▃▂    ▂▄▄▄▃▁                                             ▂
  ███████▇▅▇███████▇▇██▇▇▆▆▆▆▆▁▆▆▅▅▆▄▅▅▆▅▃▃▁▃▄▄▄▃▅▁▄▃▁▅▃▃▁▅▅▅▆ █
  50.7 μs       Histogram: log(frequency) by time      65.1 μs <

 Memory estimate: 3.33 KiB, allocs estimate: 34.
```
Which is about ~3.2x faster.