@heltonmc 2022-10-10
I went ahead and removed the instances of `@evalpoly` in the code base and changed to just `evalpoly`. Depending on how the discussion goes in #113 we can change back if needed.