@heltonmc 2022-10-11
Sorry for continuing adding more commits (last one). 

I also added faster routines to compute the airy coefficients. These (I don't think) lend them self well to use `evalpoly` but they can be computed using only one cubic root evaluation (the rest are constant propagated). For the airy coefficients since alpha only appears in a single multiply it is faster to just remove that branch (`alph==0`) completely and then we can avoid a lot of code duplication which is not the case for the Bessel coefficients.

```julia
# Master branch of the airy coefficient code block
julia> @benchmark wrapper(1.0, 1.2, x) setup=(x=rand())
BenchmarkTools.Trial: 10000 samples with 916 evaluations.
 Range (min … max):  116.826 ns … 169.379 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     119.825 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   118.818 ns ±   1.865 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▄▇▅                            ▅█▃                   ▁        ▂
  ███▅▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▇█▇▄▅▄▄▅▅▃███▄▁▁▁▁▁▁▁▁▃▁▁▁▁▁▁▁▄██▇▅▅▄▄▅▅ █
  117 ns        Histogram: log(frequency) by time        123 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

# The last commit 
julia> @benchmark gl_airy(1.0, 1.2, x) setup=(x=rand())
BenchmarkTools.Trial: 10000 samples with 998 evaluations.
 Range (min … max):  17.710 ns … 22.605 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     17.763 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   17.773 ns ±  0.147 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

                      ▂▃ ▄▅ ▆█▇▁▇█▇ ▅▄ ▂▁                      
  ▂▂▁▁▁▂▂▂▂▂▃▂▃▄▃▄▅▆▅███▆██▇██████████▇██▇▄▆▅▅▃▄▄▃▄▄▄▃▄▃▃▂▃▂▂ ▄
  17.7 ns         Histogram: frequency by time        17.8 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

The final benchmarks for this function using a 500 point quadrature look like
```julia
# Master
julia> @benchmark FastGaussQuadrature.gausslaguerre_asy(500, x) setup=(x=rand()*5.0)
BenchmarkTools.Trial: 6834 samples with 1 evaluation.
 Range (min … max):  704.500 μs …  3.611 ms  ┊ GC (min … max): 0.00% … 79.78%
 Time  (median):     725.750 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   729.383 μs ± 41.619 μs  ┊ GC (mean ± σ):  0.06% ±  0.97%

          ▁ ▄▄▇▅█▇▅▅▄▆▃▂                                        
  ▄▇▇█▆▅▆█████████████████▇▆▆▇▅▆▅▆▅▅▅▅▆▄▄▄▄▃▃▂▂▂▂▂▂▂▁▁▁▂▁▁▁▁▁▁ ▄
  704 μs          Histogram: frequency by time          777 μs <

 Memory estimate: 13.66 KiB, allocs estimate: 77.

# This PR
julia> @benchmark FastGaussQuadrature.gausslaguerre_asy(500, x) setup=(x=rand()*5.0)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  163.083 μs …  3.318 ms  ┊ GC (min … max): 0.00% … 94.24%
 Time  (median):     173.500 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   174.703 μs ± 44.615 μs  ┊ GC (mean ± σ):  0.36% ±  1.34%

   ▄▃▁    ▁▄▅▄▅▅▂▃▃▃█▇▅▆▃▁▂▄▃▂▁ ▁                               
  ▅███▆▆▆▅███████████████████████▆▅▅▄▄▃▃▃▃▃▂▂▂▂▂▂▂▂▂▁▂▂▁▂▁▁▁▁▂ ▄
  163 μs          Histogram: frequency by time          196 μs <

 Memory estimate: 13.66 KiB, allocs estimate: 77.
```
So this block of code is now about 4x faster. At this point these coefficients are a fraction of the routine and a lot of the remaining runtime and allocations are coming from the calls to special functions. In the future if you want to consider using `Bessels.jl` for the calls to `besselj`, `airyai`, and `gamma` these allocations and computations could be significantly reduced.

- Michael