@heltonmc 2022-10-31
I am unfortunately away from my desktop (Intel CPU) where I did this right this second but will run later. On my my laptop...

```julia
julia> versioninfo()
Julia Version 1.8.2
Commit 36034abf260 (2022-09-29 15:21 UTC)
Platform Info:
  OS: macOS (arm64-apple-darwin21.3.0)
  CPU: 8 × Apple M1
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-13.0.1 (ORCJIT, apple-m1)
  Threads: 1 on 8 virtual cores

# Master
julia> @benchmark FastGaussQuadrature.gausslaguerre_asy(500, x) setup=(x=rand()*5.0)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  301.875 μs …  8.494 ms  ┊ GC (min … max): 0.00% … 95.74%
 Time  (median):     317.458 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   322.533 μs ± 86.432 μs  ┊ GC (mean ± σ):  0.33% ±  1.30%

    ▁   ▂▂▃▄▃▄▇██▆▄▁                                            
  ▂▆█▇▅██████████████▇▆▅▅▅▅▆█▇▇▆▇▆▆▇▇▇▆▇█▆▆▆▄▄▃▃▂▂▂▂▂▂▂▂▂▁▂▁▁▁ ▄
  302 μs          Histogram: frequency by time          357 μs <

 Memory estimate: 13.66 KiB, allocs estimate: 77.

# This PR
julia> @benchmark FastGaussQuadrature.gausslaguerre_asy(500, x) setup=(x=rand()*5.0)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  121.208 μs …  3.452 ms  ┊ GC (min … max): 0.00% … 96.08%
 Time  (median):     127.834 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   129.131 μs ± 33.610 μs  ┊ GC (mean ± σ):  0.26% ±  0.96%

      ▆▃   ▁▇▆▅▄▇▇▅▄▆█▄           ▁                             
  ▂▇▆▇██▇███████████████▅▃▃▄▄▃▅▆▆██▇▇▆▅▅▃▂▂▂▂▁▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▄
  121 μs          Histogram: frequency by time          146 μs <

 Memory estimate: 13.66 KiB, allocs estimate: 77.
```

So about a 2.5x faster on my Mac. I kinda wish I had an AMD cpu to really see what is going here... it is hard for me to say why the increase is not occurring there. There is a significantly larger improvement on my Intel CPU because of better vectorization of the polynomial evaluations....